+++ osu.Game/Screens/Menu/MainMenu.cs
@@ -74,7 +74,7 @@ namespace osu.Game.Screens.Menu
         private SongTicker songTicker;
 
         [BackgroundDependencyLoader(true)]
-        private void load(MfMenuOverlay mfoverlay,BeatmapListingOverlay beatmapListing,DirectOverlay direct, SettingsOverlay settings, RankingsOverlay rankings, OsuConfigManager config, SessionStatics statics)
+        private void load(MfMenuOverlay mfoverlay,BeatmapListingOverlay beatmapListing, SettingsOverlay settings, RankingsOverlay rankings, OsuConfigManager config, SessionStatics statics)
         {
             holdDelay = config.GetBindable<float>(OsuSetting.UIHoldActivationDelay);
             loginDisplayed = statics.GetBindable<bool>(Static.LoginOverlayDisplayed);
@@ -136,7 +136,6 @@ namespace osu.Game.Screens.Menu
             };
 
             buttons.OnSettings = () => settings?.ToggleVisibility();
-            buttons.OnDirect = () => direct?.ToggleVisibility();
             buttons.OnChart = () => rankings?.ShowSpotlights();
             buttons.OnBeatmapListing = () => beatmapListing?.ToggleVisibility();
             buttons.OnMfMenuButton = () => mfoverlay?.ToggleVisibility();